Feat : component 분리
